Demonstrate staging db dump
`kubectl run ... --env=NAME=VALUE` interprets `$$` in VALUE as a (broken) interpolation of variables.

Values containing `$$` need to be escaped as `$$$$` as `$$` is turned
into `$`.

`sed` does this job, but specifies `$` as `\$`; `$` on its own means
end-of-the-string.

Since the `sed` shell command is inside the Makefile, all `$` further
double in number to avoid interpolation as a Make variable.

See https://github.com/kubernetes/kubernetes/issues/101137