Revert "Allow container names to be generated from project name 2531"
This reverts commit 38a5db9dc6dd260c2970864f25f9fef79e45d768.

Requires docker-compose v2 to avoid:
```
'name' does not match any of the regexes: '^x-'

You might be seeing this error because you're using the wrong Compose
file version. Either specify a supported version (e.g "2.2" or "3.3")
and place your service definitions under the `services` key, or omit the
`version` key and place your service definitions at the root of the file
to use version 1.
For more on the Compose file format versions, see
https://docs.docker.com/compose/compose-file/
```