Prepare to upgrade backstop by running backstop as root inside the container #3727
The latest of version of backstop fails to connect to the browser
session when run inside of GH Actions if --user flag is being used the
way we use it.

By only removing this when running inside of GH Actions we preserve the
existing local dev experience that files created by running backstop are
owned by the user instead of root.